refactor(topic_entity_tag): moved utils functions into separate file
